refactor(manufacture): move IngredientRequirement and ConsumptionItem to ManufactureWorkItems.cs
