+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/FlexiManufactureClient.cs
@@ -14,26 +14,6 @@ using Rem.FlexiBeeSDK.Model.Products.StockMovement;
 
 namespace Anela.Heblo.Adapters.Flexi.Manufacture;
 
-internal sealed class IngredientRequirement
-{
-    public required string ProductCode { get; init; }
-    public required string ProductName { get; init; }
-    public required ProductType ProductType { get; init; }
-    public required double RequiredAmount { get; init; }
-    public required bool HasLots { get; init; }
-}
-
-internal sealed class ConsumptionItem
-{
-    public required string ProductCode { get; init; }
-    public required string ProductName { get; init; }
-    public required ProductType ProductType { get; init; }
-    public string? LotNumber { get; init; }
-    public DateOnly? Expiration { get; init; }
-    public required double Amount { get; init; }
-    public required string SourceProductCode { get; init; }
-}
-
 internal sealed class FlexiManufactureClient : IManufactureClient
 {
     private const string WarehouseDocumentType_OutboundMaterial = "V-VYDEJ-MATERIAL";
